+++ probe-rs/src/flashing/loader.rs
@@ -1,280 +1,19 @@
-use ihex::Record;
 use probe_rs_target::{
     InstructionSet, MemoryRange, MemoryRegion, NvmRegion, RawFlashAlgorithm,
     TargetDescriptionSource,
 };
 use std::collections::HashMap;
-use std::io::{Read, Seek, SeekFrom};
+use std::io::{Read, Seek};
 use std::ops::Range;
-use std::str::FromStr;
 
 use super::builder::FlashBuilder;
-use super::{
-    extract_from_elf, BinOptions, DownloadOptions, FileDownloadError, FlashError, Flasher,
-};
-use crate::flashing::{FlashLayout, FlashProgress, FormatKind};
+use super::{DownloadOptions, FileDownloadError, FlashError, Flasher};
+use crate::flashing::image::Format;
+use crate::flashing::{FlashLayout, FlashProgress};
 use crate::memory::MemoryInterface;
 use crate::session::Session;
 use crate::Target;
 
-/// Operations related to a vendor-specific image format.
-pub trait ImageFormatDefinition: FormatClone + Send + Sync {
-    /// Returns the name of the image format.
-    fn name(&self) -> &'static str;
-
-    /// Returns a default image loader for this format.
-    fn default(&self) -> Box<dyn ImageLoader>;
-}
-
-#[doc(hidden)]
-pub trait FormatClone {
-    fn clone_box(&self) -> Box<dyn ImageFormatDefinition>;
-}
-
-impl<T> FormatClone for T
-where
-    T: 'static + ImageFormatDefinition + Clone,
-{
-    fn clone_box(&self) -> Box<dyn ImageFormatDefinition> {
-        Box::new(self.clone())
-    }
-}
-
-impl Clone for Box<dyn ImageFormatDefinition> {
-    fn clone(&self) -> Self {
-        self.clone_box()
-    }
-}
-
-/// A handler for a vendor-specific image format.
-#[derive(Clone)]
-pub struct ImageFormat(Box<dyn ImageFormatDefinition>);
-
-impl ImageFormat {
-    /// Returns the name of the image format.
-    pub fn name(&self) -> &'static str {
-        self.0.name()
-    }
-
-    /// Returns a default image loader for this format.
-    pub fn default(&self) -> Box<dyn ImageLoader> {
-        self.0.default()
-    }
-}
-
-impl PartialEq for ImageFormat {
-    fn eq(&self, other: &Self) -> bool {
-        self.name() == other.name()
-    }
-}
-impl Eq for ImageFormat {}
-
-impl std::fmt::Debug for ImageFormat {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(self.name())
-    }
-}
-
-impl std::fmt::Display for ImageFormat {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(self.name())
-    }
-}
-
-impl FromStr for ImageFormat {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        crate::vendor::try_parse_image_format(s)
-            .ok_or_else(|| format!("Unknown image format '{s}'"))
-    }
-}
-
-/// Helper trait for object safety.
-pub trait ImageReader: Read + Seek {}
-impl<T> ImageReader for T where T: Read + Seek {}
-
-/// Load and parse a firmware in a particular format, and add it to the flash loader.
-///
-/// Based on the image loader, probe-rs may apply certain transformations to the firmware.
-pub trait ImageLoader {
-    /// Loads the given image.
-    fn load(
-        &self,
-        flash_loader: &mut FlashLoader,
-        session: &mut Session,
-        file: &mut dyn ImageReader,
-    ) -> Result<(), FileDownloadError>;
-}
-
-/// An initialized firmware image loader.
-pub struct Format(Box<dyn ImageLoader>);
-
-impl From<FormatKind> for Format {
-    fn from(kind: FormatKind) -> Self {
-        match kind {
-            FormatKind::Bin => Self::from(BinLoader(BinOptions::default())),
-            FormatKind::Hex => Self::from(HexLoader),
-            FormatKind::Elf => Self::from(ElfLoader),
-            FormatKind::Uf2 => Self::from(Uf2Loader),
-            FormatKind::VendorSpecific(f) => Self(f.default()),
-        }
-    }
-}
-
-impl<I> From<I> for Format
-where
-    I: ImageLoader + 'static,
-{
-    fn from(loader: I) -> Self {
-        Self(Box::new(loader))
-    }
-}
-
-/// Reads the data from the binary file and adds it to the loader without splitting it into flash instructions yet.
-pub struct BinLoader(pub BinOptions);
-
-impl ImageLoader for BinLoader {
-    fn load(
-        &self,
-        flash_loader: &mut FlashLoader,
-        _session: &mut Session,
-        file: &mut dyn ImageReader,
-    ) -> Result<(), FileDownloadError> {
-        // Skip the specified bytes.
-        file.seek(SeekFrom::Start(u64::from(self.0.skip)))?;
-
-        let mut buf = Vec::new();
-        file.read_to_end(&mut buf)?;
-
-        flash_loader.add_data(
-            // If no base address is specified use the start of the boot memory.
-            // TODO: Implement this as soon as we know targets.
-            self.0.base_address.unwrap_or_default(),
-            &buf,
-        )?;
-
-        Ok(())
-    }
-}
-
-/// Prepares the data sections that have to be loaded into flash from an ELF file.
-/// This will validate the ELF file and transform all its data into sections but no flash loader commands yet.
-pub struct ElfLoader;
-
-impl ImageLoader for ElfLoader {
-    fn load(
-        &self,
-        flash_loader: &mut FlashLoader,
-        _session: &mut Session,
-        file: &mut dyn ImageReader,
-    ) -> Result<(), FileDownloadError> {
-        let mut elf_buffer = Vec::new();
-        file.read_to_end(&mut elf_buffer)?;
-
-        let extracted_data = extract_from_elf(&elf_buffer)?;
-
-        if extracted_data.is_empty() {
-            tracing::warn!("No loadable segments were found in the ELF file.");
-            return Err(FileDownloadError::NoLoadableSegments);
-        }
-
-        tracing::info!("Found {} loadable sections:", extracted_data.len());
-
-        for section in &extracted_data {
-            let source = match section.section_names.len() {
-                0 => "Unknown",
-                1 => section.section_names[0].as_str(),
-                _ => "Multiple sections",
-            };
-
-            tracing::info!(
-                "    {} at {:#010X} ({} byte{})",
-                source,
-                section.address,
-                section.data.len(),
-                if section.data.len() == 1 { "" } else { "s" }
-            );
-        }
-
-        for data in extracted_data {
-            flash_loader.add_data(data.address.into(), data.data)?;
-        }
-
-        Ok(())
-    }
-}
-
-/// Reads the HEX data segments and adds them as loadable data blocks to the loader.
-/// This does not create any flash loader instructions yet.
-pub struct HexLoader;
-
-impl ImageLoader for HexLoader {
-    fn load(
-        &self,
-        flash_loader: &mut FlashLoader,
-        _session: &mut Session,
-        file: &mut dyn ImageReader,
-    ) -> Result<(), FileDownloadError> {
-        let mut base_address = 0;
-
-        let mut data = String::new();
-        file.read_to_string(&mut data)?;
-
-        for record in ihex::Reader::new(&data) {
-            match record? {
-                Record::Data { offset, value } => {
-                    let offset = base_address + offset as u64;
-                    flash_loader.add_data(offset, &value)?;
-                }
-                Record::ExtendedSegmentAddress(address) => {
-                    base_address = (address as u64) * 16;
-                }
-                Record::ExtendedLinearAddress(address) => {
-                    base_address = (address as u64) << 16;
-                }
-
-                Record::EndOfFile
-                | Record::StartSegmentAddress { .. }
-                | Record::StartLinearAddress(_) => {}
-            }
-        }
-        Ok(())
-    }
-}
-
-/// Prepares the data sections that have to be loaded into flash from an UF2 file.
-/// This will validate the UF2 file and transform all its data into sections but no flash loader commands yet.
-pub struct Uf2Loader;
-
-impl ImageLoader for Uf2Loader {
-    fn load(
-        &self,
-        flash_loader: &mut FlashLoader,
-        _session: &mut Session,
-        file: &mut dyn ImageReader,
-    ) -> Result<(), FileDownloadError> {
-        let mut uf2_buffer = Vec::new();
-        file.read_to_end(&mut uf2_buffer)?;
-
-        let (converted, family_to_target) = uf2_decode::convert_from_uf2(&uf2_buffer).unwrap();
-        let target_addresses = family_to_target.values();
-        let num_sections = family_to_target.len();
-
-        if let Some(target_address) = target_addresses.min() {
-            tracing::info!("Found {} loadable sections:", num_sections);
-            if num_sections > 1 {
-                tracing::warn!("More than 1 section found in UF2 file.  Using first section.");
-            }
-            flash_loader.add_data(*target_address, &converted)?;
-
-            Ok(())
-        } else {
-            tracing::warn!("No loadable segments were found in the UF2 file.");
-            Err(FileDownloadError::NoLoadableSegments)
-        }
-    }
-}
 /// `FlashLoader` is a struct which manages the flashing of any chunks of data onto any sections of flash.
 ///
 /// Use [add_data()](FlashLoader::add_data) to add a chunk of data.
@@ -372,7 +111,7 @@ impl FlashLoader {
             }
         }
 
-        loader.0.load(self, session, file)
+        loader.load(self, session, file)
     }
 
     /// Writes all the stored data chunks to flash.
